+++ src/platform.ts
@@ -1,5 +1,5 @@
-import { ResolvedVersion } from './version';
-import { execPrivileged, execPrivilegedWithOutput } from './utils';
+import { ResolvedVersion } from './version.js';
+import { execPrivileged, execPrivilegedWithOutput } from './utils.js';
 
 export interface PackageManager {
     accept<T>(v: PackageManagerVisitor<T>): T;
@@ -145,7 +145,7 @@ export class Yum extends Dnf implements PackageManager {
 export class Zypper implements PackageManager {
     // This package is part of the `--plus-content debug` repository
     private readonly debugInfoPackages: string[] = ['glibc-debuginfo'];
-    private readonly valgrindBuildDeps: string[] = ['gcc', 'make', 'bzip2'];
+    private readonly valgrindBuildDeps: string[] = ['gcc', 'make', 'bzip2', 'findutils'];
 
     accept<T>(v: PackageManagerVisitor<T>) {
         return v.visitZypper(this);